+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added a `scrollableAxesSafeAreaEdges` property to `ScrollView` that is used with `ContentInsetAdjustmentBehavior.scrollableAxes`.
+
 ### Removed
 
 ### Changed